requirements.txt
langchain
langchain-openai
langchain-chroma
chromadb
openai
python-dotenv
fastapi
uvicorn
